break up processed rna seq files etl
