+++ polyphemus/lib/etls.rb
@@ -12,4 +12,7 @@ require_relative 'etls/sftp_ingest_cat_to_c4_triage_files_etl'
 require_relative 'etls/ipi_rna_seq_link_fastqs_etl'
 require_relative 'etls/ipi_rna_seq_populate_attributes_etl'
 require_relative 'etls/ipi_rna_seq_populate_matrices_etl'
-require_relative 'etls/ipi_rna_seq_link_processed_files_etl'
+require_relative 'etls/ipi_rna_seq_link_cram_files_etl'
+require_relative 'etls/ipi_rna_seq_link_cram_index_files_etl'
+require_relative 'etls/ipi_rna_seq_link_junction_files_etl'
+require_relative 'etls/ipi_rna_seq_link_unmapped_fastq_files_etl'
